Fix a bug that Flyway migrations are not run into Importer's own internal database.
Make a change that Flyway migrations are always run into Importer's own internal
database regardless of whether it is a development or production environment.
Previously, the Flyway configuration was such that the migrations targeted
the Jore4 database, but this is no longer the case. Now we want to keep
the internal database of Importer always up to date.