Create PostgreSQL extensions conditionally.
This is because in Azure PostgreSQL Flexible Server v15, ordinary users (without
the azure_pg_admin role) cannot create extensions. These extensions need to be
created before deploying this service into the cloud.

Add the creation of the PostGIS extension to the migrations for consistency and
because it may not be available in all database environments.

When using PostGIS Docker image as a base for database, the postgis extension
already exists.